Add tool invocation wrapper
