+++ src/meta_mcp/tooling/__init__.py
@@ -1,0 +1,2 @@
+"""Tooling helpers for Meta MCP."""
+
